+++ docs/changelog.rst
@@ -9,6 +9,9 @@ track version numbers, where backwards incompatible changes
 latest changes in development for next release
 ----------------------------------------------
 
+* **standardized encoding of output with ``-e/--encoding`` option**
+  (`#39`_)
+
 * support for ``.xls`` and ``.xlsx`` files (`#42`_ and `#55`_ by `@levivm`_)
 
 * support for ``.epub`` files (`#40`_ by `@kokxx`_)
@@ -22,6 +25,11 @@ latest changes in development for next release
 
   * concurrancy problems with tesseract (`#44`_ by `@ShawnMilo`_,
     `#41`_ by `@christomitov`_)
+
+* several internal improvements, including:
+
+  * switching to using class-based parsers to abstract away the common
+    functionality between different parser classes (`#39`_)
 
 
 0.5.1
@@ -126,6 +134,7 @@ latest changes in development for next release
 .. _#29: https://github.com/deanmalmgren/textract/issues/29
 .. _#30: https://github.com/deanmalmgren/textract/issues/30
 .. _#33: https://github.com/deanmalmgren/textract/issues/33
+.. _#39: https://github.com/deanmalmgren/textract/issues/39
 .. _#40: https://github.com/deanmalmgren/textract/issues/40
 .. _#41: https://github.com/deanmalmgren/textract/issues/41
 .. _#42: https://github.com/deanmalmgren/textract/issues/42